backend/requirements.txt
# API
gunicorn==21.2.0
uvicorn==0.23.2
fastapi==0.104.0
pydantic-settings==2.0.3
pydantic[email]
orjson==3.9.9
passlib[bcrypt]
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
websockets
# БД
fastapi-users[sqlalchemy]
SQLAlchemy==2.0.22
asyncpg==0.28.0
alembic==1.12.0
psycopg2-binary==2.9.9
redis==5.0.1